data_generator/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asn1crypto==1.5.1
attrs==25.1.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloud-sql-python-connector==1.17.0
cryptography==44.0.1
dnspython==2.7.0
Faker==33.3.0
frozenlist==1.5.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-logging==3.11.4
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
idna==3.10
multidict==6.1.0
pg8000==1.31.2
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
scramp==1.4.5
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
yarl==1.18.3
